Initial and working version of the application
This version already works, thus object files can be modified, then
linked with the remaining parts. The resulting application works in all
tested cases, but some corner cases may not work right now.
Also the code quality might not be the best ;)

The README-GER.txt contains a more detailed description of how the
application works, but only in German! An explanation in English will be
available later in the README.md